Phase 2: HTTP layer extraction complete
- Created serv/http/requests.py with request classes and method mapping
- Created serv/http/responses.py with all response classes and ResponseBuilder
- Created serv/http/forms.py with form processing and validation utilities
- Moved request classes: GetRequest, PostRequest, etc. from routes.py
- Moved response classes: JsonResponse, TextResponse, etc. from routes.py
- Moved ResponseBuilder and protocols from responses.py
- Moved form utilities: Form class, validators, FileUpload from routes.py and requests.py
- Updated all imports to use new locations
- Maintained backward compatibility through re-exports
- Updated serv/http/__init__.py to provide clean API
- Old import paths still work, new organized paths available
- Tests continue to pass
