+++ serv/http/__init__.py
@@ -2,15 +2,16 @@
 
 # Re-export from existing modules for backward compatibility
 try:
-    from serv.routes import GetRequest, PostRequest, PutRequest, DeleteRequest, PatchRequest
-    from serv.responses import (
-        JSONResponse,
-        HTMLResponse, 
+    from .requests import GetRequest, PostRequest, PutRequest, DeleteRequest, PatchRequest
+    from .responses import (
+        JsonResponse as JSONResponse,  # Alias for backward compatibility
+        HtmlResponse as HTMLResponse,  # Alias for backward compatibility
         TextResponse,
         RedirectResponse,
         Response,
         ResponseBuilder,
     )
+    from .forms import FileUpload, Form
     
     __all__ = [
         "GetRequest",
@@ -24,6 +25,8 @@ try:
         "RedirectResponse",
         "Response",
         "ResponseBuilder",
+        "FileUpload",
+        "Form",
     ]
 except ImportError:
     # If direct imports fail, will be populated later during refactoring
